train: wave-D assembly — visui Align one-liner (ISS-770 ripple), 13 covenant re-stamps (SgeError, Music, Font, TextraLabel, TypingLabel, GLTFExtras +entries+keys, SceneManager -4 accessors, Scene, AnimationControllerHack, ModelInstanceHack +deviation note, GLTFJsonParser, PBREmissiveShaderProvider, GLTFCodecs ISS-629 reconciliation), 3 skip-policy entries (gltf test fixtures + restored upstream XXX comment)

+++ sge-extension/visui/src/main/scala/sge/visui/widget/VisTextArea.scala
@@ -252,7 +252,7 @@ class VisTextArea(text: Nullable[String], visStyle: VisTextField.VisTextFieldSty
     var offsetY = 0f
     var i       = firstLineShowing * 2
     while (i < (firstLineShowing + linesShowing) * 2 && i < linesBreak.size) {
-      font.draw(batch, displayText, x, y + offsetY, linesBreak.items(i), linesBreak.items(i + 1), 0, Align.left.toInt, false)
+      font.draw(batch, displayText, x, y + offsetY, linesBreak.items(i), linesBreak.items(i + 1), 0, Align.left, false)
       offsetY -= font.lineHeight
       i += 2
     }
